Primera fase Completada. Corregir funcion temperatura.

+++ SistemaRiego/src/Headers/sensors.h
@@ -58,7 +58,7 @@ uint8_t getlvlH2O(uint8_t lvl);
 int getSensorValue(int reg);
 int convertToTemp(int n);
 uint8_t convertToLvlH2O(uint16_t n);
-uint8_t convertToHum(uint16_t n);
+uint8_t convertToHum(int n);
